separating daq monitoring functions from amc functions and providing dedicated files
