+++ Makefile
@@ -27,6 +27,7 @@ TARGET_LIBS=lib/memory.so
 TARGET_LIBS+=lib/optical.so
 TARGET_LIBS+=lib/utils.so
 TARGET_LIBS+=lib/extras.so
+TARGET_LIBS+=lib/daq_monitor.so
 TARGET_LIBS+=lib/amc.so
 TARGET_LIBS+=lib/vfat3.so
 TARGET_LIBS+=lib/optohybrid.so
@@ -45,6 +46,9 @@ lib/utils.so: src/utils.cpp
 
 lib/extras.so: src/extras.cpp
 	$(CXX) $(CFLAGS) -std=c++1y -O3 -pthread $(INC) $(LDFLAGS) -fPIC -shared -Wl,-soname,extras.so -o $@ $< -lwisci2c -lxhal -llmdb
+
+lib/daq_monitor.so: src/daq_monitor.cpp
+	$(CXX) $(CFLAGS) -std=c++1y -O3 -pthread $(INC) $(LDFLAGS) -fPIC -shared -Wl,-soname,daq_monitor.so -o $@ $< -lwisci2c -lxhal -llmdb -l:utils.so -l:extras.so
 
 lib/amc.so: src/amc.cpp
 	$(CXX) $(CFLAGS) -std=c++1y -O3 -pthread $(INC) $(LDFLAGS) -fPIC -shared -Wl,-soname,amc.so -o $@ $< -lwisci2c -lxhal -llmdb -l:utils.so -l:extras.so
